refactor(ai-diagnostic): extract transport + HTTP adapters from the monolith
Continues the diagnostic_service split, test-first. Added 14 characterisation
tests over the adapter layer — transport adapters via fake sockets (TCP parse->
dispatch->serialise, Serial framing) and the HTTPAdapter Flask surface via a
real test client (health, diagnose, HMAC auth 401/403/200, learning stats) —
then extracted:

  service/transport_adapters.py - ProtocolAdapter, TCP/Serial/MQTT adapters
  service/optional_services.py  - the ~14 optional AI-subsystem imports, their
                                  *_AVAILABLE flags and the ERR_*/ROUTE_ constants,
                                  shared so the split avoids a circular import
  service/http_adapter.py       - HTTPAdapter (Flask API, ~50 routes)

diagnostic_service.py re-exports every name, so importers and the runtime
entrypoint are unchanged. The monolith is now 2807 -> 386 LOC (an 86% cut;
orchestration + main only), split across 8 cohesive modules. All 187
ai-diagnostic tests pass (153 original + 34 new characterisation tests).
